+++ user/templates/user/password.html
@@ -1,5 +1,5 @@
 {% load static %}
-<link rel="stylesheet" type="text/css" href="{% static 'signup/style.css' %}">
+<link rel="stylesheet" type="text/css" href="{% static 'user/style.css' %}">
 
 <body>
   <h2>You'll need a password</h2>
