+++ user/templates/user/signup.html
@@ -1,5 +1,5 @@
 {% load static %}
-<link rel="stylesheet" type="text/css" href="{% static 'signup/style.css' %}">
+<link rel="stylesheet" type="text/css" href="{% static 'user/style.css' %}">
 
 <body>
   <h2>Create your account</h2>
