+++ user/urls.py
@@ -2,7 +2,7 @@ from django.urls import path
 from . import views
 
 
-app_name = 'signup'
+app_name = 'user'
 urlpatterns = [
     path('', views.signupPage, name="signup"),
     path('password', views.passwordPage, name="create_password"),
